Chia lại các hàn lquan đến account và customer trong AuthenticationController sang AccountController và CustomerController
