+++ GHCW_BE/Models/Schedule.cs
@@ -9,7 +9,6 @@ namespace GHCW_BE.Models
         public int ReceptionistId { get; set; }
         public byte Shift { get; set; }
         public DateTime Date { get; set; }
-        public bool IsActive { get; set; }
 
         public virtual Account Receptionist { get; set; } = null!;
     }
