+++ GHCW_BE/Services/AuthenticationService.cs
@@ -15,8 +15,8 @@ namespace GHCW_BE.Services
     {
         private readonly GHCWContext _context;
         private readonly IConfiguration _configuration;
-        private Helper _helper;
-        private IMapper _mapper;
+        private readonly Helper _helper;
+        private readonly IMapper _mapper;
 
         public AuthenticationService(GHCWContext context, IConfiguration configuration, Helper helper, IMapper mapper)
         {
@@ -32,45 +32,11 @@ namespace GHCW_BE.Services
             await _context.SaveChangesAsync();
         }
 
-        //kiem tra nguoi dung co ton tai hay khong
-        //neu co tra ve account, neu khong tra ve null
-        public async Task<Account?> CheckAccountExsit(string email)
+        public string CreateValidationLink(string encodedEmail, string encodedActivationCode)
         {
-            var checkUser = await _context.Accounts.FirstOrDefaultAsync(x => x.Email.Equals(email));
-            return checkUser;
-        }
-        public async Task<CustomerDTO?> CheckCustomerExsit(int aId)
-        {
-            var checkCustomer = await _context.Customers.FirstOrDefaultAsync(x => x.AccountId == aId);
-            if (checkCustomer == null)
-            {
-                return null;
-            }
-            var customerDTO = _mapper.Map<Customer, CustomerDTO>(checkCustomer);
-            return customerDTO;
-        }
-        public async Task<CustomerDTO?> CheckPhoneExsit(string phone)
-        {
-            var checkCustomer = await _context.Customers.FirstOrDefaultAsync(x => x.PhoneNumber.Equals(phone));
-            if (checkCustomer == null)
-            {
-                return null;
-            }
-            var customerDTO = _mapper.Map<Customer, CustomerDTO>(checkCustomer);
-            return customerDTO;
-        }
-
-        //kiem tra nguoi dung co ton tai va dang hoat dong hay khong
-        //neu co tra ve account, neu khong tra ve null
-        public async Task<AccountDTO?> CheckActiveStatus(string email)
-        {
-            var existUser = await _context.Accounts.FirstOrDefaultAsync(u => u.Email == email && u.IsActive == true);
-            if (existUser != null)
-            {
-                var existUserDTO = _mapper.Map<Account, AccountDTO>(existUser);
-                return existUserDTO;
-            }
-            return null;
+            var emailSettings = _configuration.GetSection("JWT")["Audience"];
+            var activationLink = $"{emailSettings}/Authentications/EmailActivation?email={encodedEmail}&code={encodedActivationCode}";
+            return activationLink;
         }
 
         //Kich hoat tai khoan, sau khi kich hoat se xoa code
@@ -182,67 +148,6 @@ namespace GHCW_BE.Services
             }
         }
 
-        //lấy thông tin user bằng ID
-        public async Task<AccountDTO?> GetUserProfileById(int uID)
-        {
-            var user = await _context.Accounts.FirstOrDefaultAsync(u => u.Id == uID);
-            if (user != null)
-            {
-                var userDTO = _mapper.Map<Account, AccountDTO>(user);
-                return userDTO;
-            }
-            return null;
-        }
-
-        //lấy danh sách user
-        public async Task<List<AccountDTO>?> GetUserList()
-        {
-            var users = await _context.Accounts.ToListAsync();
-            if (users != null)
-            {
-                var userDTOs = _mapper.Map<List<Account>, List<AccountDTO>>(users);
-                return userDTOs;
-            }
-            return null;
-        }
-
-        //lấy danh sách nhân viên role từ 1 - 4
-        public async Task<List<AccountDTO>?> GetEmployeeList()
-        {
-            var employees = await _context.Accounts.Where(a => a.Role >= 1 && a.Role <= 4).ToListAsync();
-            if (employees != null)
-            {
-                var employeeDTOs = _mapper.Map<List<Account>, List<AccountDTO>>(employees);
-                return employeeDTOs;
-            }
-            return null;
-        }
-
-        //lấy danh sách lễ tân role = 4
-        public async Task<List<AccountDTO>?> GetReceptionList()
-        {
-            var employees = await _context.Accounts.Where(a => a.Role == 4).ToListAsync();
-            if (employees != null)
-            {
-                var employeeDTOs = _mapper.Map<List<Account>, List<AccountDTO>>(employees);
-                return employeeDTOs;
-            }
-            return null;
-        }
-
-        //lấy danh sách khách hàng role = 5
-        public async Task<List<AccountDTO>?> GetCustomerAccountList()
-        {
-            var customers = await _context.Accounts.Where(a => a.Role == 5).ToListAsync();
-            if (customers != null)
-            {
-                var customerDTOs = _mapper.Map<List<Account>, List<AccountDTO>>(customers);
-                return customerDTOs;
-            }
-            return null;
-        }
-
-        //Đổi password
         public async Task<bool> ChangePassword(int uId, string newPassword)
         {
             var user = await _context.Accounts.FindAsync(uId);
@@ -259,8 +164,12 @@ namespace GHCW_BE.Services
         public async Task<AccountDTO?> GetAccountByRefreshToken(string refreshToken)
         {
             var acc = await _context.Accounts.FirstOrDefaultAsync(a => a.RefreshToken == refreshToken);
-            var accDTO = _mapper.Map<Account, AccountDTO>(acc);
-            return accDTO;
+            if (acc != null)
+            {
+                var accDTO = _mapper.Map<Account, AccountDTO>(acc);
+                return accDTO;
+            }
+            return null;
         }
 
         public async Task DeleteRefToken(AccountDTO a)
@@ -273,6 +182,7 @@ namespace GHCW_BE.Services
                 await _context.SaveChangesAsync();
             }
         }
+
         public async Task<(bool isSuccess, string message)> UpdateRefreshToken(AccountDTO a)
         {
             var user = await _context.Accounts.FirstOrDefaultAsync(u => u.Id == a.Id);
@@ -291,142 +201,6 @@ namespace GHCW_BE.Services
             {
                 return (false, "Cập nhật thông tin thất bại, vui lòng kiểm tra lại.");
             }
-        }
-        public async Task<(bool isSuccess, string message)> UserActivation(int uid)
-        {
-            var acc = await _context.Accounts.FindAsync(uid);
-            if (acc == null)
-            {
-                return (false, "Tài khoản không tồn tại.");
-            }
-            try
-            {
-                acc.IsActive = !acc.IsActive;
-                _context.Accounts.Update(acc);
-                await _context.SaveChangesAsync();
-
-                return (true, "Thay đổi trạng thái tài khoản thành công.");
-            }   
-            catch (Exception)
-            {
-                return (false, "Thay đổi trạng thái tài khoản thất bại, vui lòng thử lại.");
-            }
-        }
-
-        public async Task<(bool isSuccess, string message)> AddNewUser(AddRequest a)
-        {
-            try
-            {
-                var user = _mapper.Map<AddRequest, Account>(a);
-                _context.Accounts.Add(user);
-                await _context.SaveChangesAsync();
-                return (true, "Thêm tài khoản mới thành công.");
-            }
-            catch (Exception)
-            {
-                return (false, "Có lỗi trong quá trình thêm tài khoản, vui lòng thử lại.");
-            }
-        }
-
-        //user tự chỉnh sửa profile của mình
-        public async Task<(bool isSuccess, string message)> UpdateProfile(UpdateRequest r)
-        {
-            var user = await _context.Accounts.FirstOrDefaultAsync(u => u.Id == r.Id);
-            if (user == null)
-            {
-                return (false, "Không tìm thấy tài khoản.");
-            }
-            try
-            {
-                _mapper.Map(r, user);
-                _context.Accounts.Update(user);
-                await _context.SaveChangesAsync();
-                return (true, "Cập nhật thông tin thành công.");
-            }
-            catch (Exception)
-            {
-                return (false, "Cập nhật thông tin thất bại, vui lòng kiểm tra lại.");
-            }
-        }
-
-        //admin chỉnh sửa thông tin của user
-        public async Task<(bool isSuccess, string message)> EditProfile(EditRequest r)
-        {
-            var user = await _context.Accounts.FirstOrDefaultAsync(u => u.Id == r.Id);
-            if (user == null)
-            {
-                return (false, "Không tìm thấy tài khoản.");
-            }
-            try
-            {
-                _mapper.Map(r, user);
-                _context.Accounts.Update(user);
-                await _context.SaveChangesAsync();
-                return (true, "Cập nhật thông tin thành công.");
-            }
-            catch (Exception)
-            {
-                return (false, "Cập nhật thông tin thất bại, vui lòng kiểm tra lại.");
-            }
-        }
-
-        public async Task<List<CustomerDTO>> GetCustomerList()
-        {
-            var customers = await _context.Customers.Include(c => c.Account).ToListAsync();
-            var customerDTOs = _mapper.Map<List<Customer>, List<CustomerDTO>>(customers);
-            return customerDTOs;
-        }
-
-        public async Task<CustomerDTO?> GetCustomerProfileById(int uID)
-        {
-            var customer = await _context.Customers.Include(c => c.Account).FirstOrDefaultAsync(u => u.Id == uID);
-            if (customer != null)
-            {
-                var userDTO = _mapper.Map<Customer, CustomerDTO>(customer);
-                return userDTO;
-            }
-            return null;
-        }
-
-        public async Task<(bool isSuccess, string message)> AddNewCustomer(AddCustomerRequest a)
-        {
-            try
-            {
-                var customer = _mapper.Map<AddCustomerRequest, Customer>(a);
-                _context.Customers.Add(customer);
-                await _context.SaveChangesAsync();
-                return (true, "Thêm khách hàng mới thành công.");
-            }
-            catch (Exception)
-            {
-                return (false, "Có lỗi trong quá trình thêm khách hàng mới, vui lòng thử lại.");
-            }
-        }
-
-        public async Task<(bool isSuccess, string message)> EditCustomer(CustomerDTO c)
-        {
-            var customer = await _context.Customers.FirstOrDefaultAsync(u => u.Id == c.Id);
-            if (customer == null)
-            {
-                return (false, "Không tìm thấy khách hàng.");
-            }
-            try
-            {
-                _mapper.Map(c, customer);
-                _context.Customers.Update(customer);
-                await _context.SaveChangesAsync();
-                return (true, "Cập nhật thông tin thành công.");
-            }
-            catch (Exception)
-            {
-                return (false, "Cập nhật thông tin thất bại, vui lòng kiểm tra lại.");
-            }
-        }
-
-        public async Task<IEnumerable<Ticket?>> GetBookingListById(int uid)
-        {
-            var tickets = await _context.Tickets.Where(t => t.CustomerId == uid).ToListAsync();
-            return tickets;
         }
     }
 }
