+++ GHCW_BE/appsettings.json
@@ -16,7 +16,7 @@
     "RefSecretKey": "truongkikhangchiennhatdinhthangloi",
     "Audience": "http://localhost:7260",
     "Issuer": "http://localhost:7226",
-    //"Audience": "https://geothermalhotspringcenterwebsite.azurewebsites.net/",
+    //"Audience": "https://geothermalhotspringcenterwebsite.azurewebsites.net",
     //"Issuer": "http://ghcw.azurewebsites.net",
     "Expiration": "30",
     "RefExpiration": "1324800" //khoảng 3 tháng
