+++ GHCW_FE/Pages/Admin/CustomerManagement.cshtml
@@ -42,8 +42,8 @@
 
         <select name="SortOption" onchange="document.getElementById('filterForm').submit()">
             <option value="0" @(Model.SortOption == 0 ? "selected" : "")>Lọc theo</option>
-            <option value="1" @(Model.SortOption == 1 ? "selected" : "")>A-Z</option>
-            <option value="2" @(Model.SortOption == 2 ? "selected" : "")>Z-A</option>
+            <option value="1" @(Model.SortOption == 1 ? "selected" : "")>Khách đã đăng ký tài khoản</option>
+            <option value="2" @(Model.SortOption == 2 ? "selected" : "")>Khách chưa đăng ký tài khoản</option>
             <option value="3" @(Model.SortOption == 3 ? "selected" : "")>Khách nam</option>
             <option value="4" @(Model.SortOption == 4 ? "selected" : "")>Khách nữ</option>
         </select>
