+++ GHCW_FE/Pages/Admin/ScheduleManagement.cshtml
@@ -82,7 +82,7 @@
                                     <form method="post" asp-page-handler="DeleteSchedule">
                                         <input type="hidden" name="startDate2" value="@Model.SW.StartDate.ToString("yyyy-MM-dd")" />
                                         <input type="hidden" name="endDate2" value="@Model.SW.EndDate.ToString("yyyy-MM-dd")" />
-                                        <button type="submit" class="btn btn-danger" name="id" value="@scheduledEmployee.Id">Ẩn</button>
+                                        <button type="submit" class="btn btn-danger" name="id" value="@scheduledEmployee.Id">Xóa</button>
                                     </form>
                                 </div>
                             }
