+++ GHCW_FE/Pages/Authentications/Login.cshtml.cs
@@ -3,6 +3,7 @@ using GHCW_FE.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
+using System.Net;
 using System.Net.Http.Headers;
 
 namespace GHCW_FE.Pages.Authentications
@@ -10,10 +11,14 @@ namespace GHCW_FE.Pages.Authentications
     public class LoginModel : PageModel
     {
         private readonly AuthenticationService _authService;
+        private readonly TokenService _tokenService;
+        private readonly AccountService _accService;
 
-        public LoginModel(AuthenticationService authService)
+        public LoginModel(AuthenticationService authService, TokenService tokenService, AccountService accService)
         {
             _authService = authService;
+            _tokenService = tokenService;
+            _accService = accService;
         }
 
         [BindProperty]
@@ -31,10 +36,18 @@ namespace GHCW_FE.Pages.Authentications
                 return Page();
             }
 
-            if (await _authService.AuthenticateUserAsync(Account.Email, Account.Password))
+            if (await _authService.AuthenticateUserAsync(Account))
             {
-                var userAccount = await _authService.GetUserInfoAsync();
-                if (userAccount != null)
+                var accessToken = await _tokenService.CheckAndRefreshTokenAsync();
+                if (string.IsNullOrEmpty(accessToken))
+                {
+                    TempData["ErrorMessage"] = "Không lấy được token đăng nhập.";
+                    return RedirectToPage("/Authentications/Login");
+                }
+                _authService.SetAccessToken(accessToken);
+
+                var (statusCode, userAccount) = await _accService.UserProfile(accessToken);
+                if (statusCode == HttpStatusCode.OK && userAccount != null)
                 {
                     // Lưu thông tin tài khoản vào session
                     HttpContext.Session.SetString("acc", System.Text.Json.JsonSerializer.Serialize(userAccount));
@@ -52,8 +65,14 @@ namespace GHCW_FE.Pages.Authentications
                         2 => Redirect("../Index"),
                         3 => Redirect("../Index"),
                         4 => Redirect("../Index"),
-                        5 => Redirect("../Index")
+                        5 => Redirect("../Index"),
+                        _ => RedirectToPage("/Index")
                     };
+                }
+                else
+                {
+                    TempData["ErrorMessage"] = "Không thể tải thông tin người dùng. Vui lòng thử lại.";
+                    return RedirectToPage("/Authentications/Login");
                 }
             }
 
@@ -63,6 +82,14 @@ namespace GHCW_FE.Pages.Authentications
 
         public async Task<IActionResult> OnPostLogoutAsync()
         {
+            var accessToken = await _tokenService.CheckAndRefreshTokenAsync();
+            if (string.IsNullOrEmpty(accessToken))
+            {
+                TempData["ErrorMessage"] = "Không lấy được token đăng nhập.";
+                return RedirectToPage("/Authentications/Login");
+            }
+            _authService.SetAccessToken(accessToken);
+
             await _authService.LogoutAsync();
             return RedirectToPage("/Authentications/Login");
         }
